+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/items/HashedArmorpiece.java
@@ -61,8 +61,8 @@ public final class HashedArmorpiece {
             this.hash = copy.hashCode();
         }
 
-        if (item instanceof SlimefunArmorPiece) {
-            this.item = Optional.of((SlimefunArmorPiece) item);
+        if (item instanceof SlimefunArmorPiece armorPiece) {
+            this.item = Optional.of(armorPiece);
         } else {
             this.item = Optional.empty();
         }