+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/commands/subcommands/DebugFishCommand.java
@@ -19,8 +19,8 @@ class DebugFishCommand extends SubCommand {
 
     @Override
     public void onExecute(CommandSender sender, String[] args) {
-        if (sender instanceof Player && sender.hasPermission("slimefun.debugging")) {
-            ((Player) sender).getInventory().addItem(SlimefunItems.DEBUG_FISH.clone());
+        if (sender instanceof Player player && sender.hasPermission("slimefun.debugging")) {
+            player.getInventory().addItem(SlimefunItems.DEBUG_FISH.clone());
         } else {
             Slimefun.getLocalization().sendMessage(sender, "messages.no-permission", true);
         }