+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/Slimefun.java
@@ -134,6 +134,11 @@ import me.mrCookieSlime.Slimefun.api.inventory.UniversalBlockMenu;
  */
 public final class Slimefun extends JavaPlugin implements SlimefunAddon {
 
+    /**
+     * This is the Java version we recommend server owners to use.
+     * This does not necessarily mean that it's the minimum version
+     * required to run Slimefun.
+     */
     private static final int RECOMMENDED_JAVA_VERSION = 17;
 
     /**